refactor: #9 creat a mealSearch package
