+++ src/main/java/be/multimedi/mealplanning/mealSearch/MealRepository.java
@@ -1,11 +1,10 @@
-package be.multimedi.mealplanning.mealPlan;
+package be.multimedi.mealplanning.mealSearch;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface MealRepository extends JpaRepository<Meal, Long> {
     @Query("SELECT m FROM Meal m WHERE m.calories >= :minCalories AND m.calories <= :maxCalories")
