+++ src/main/java/be/multimedi/mealplanning/mealSearch/MealService.java
@@ -1,4 +1,4 @@
-package be.multimedi.mealplanning.mealPlan;
+package be.multimedi.mealplanning.mealSearch;
 
 import java.util.List;
