+++ src/main/java/be/multimedi/mealplanning/mealSearch/MealServiceImpl.java
@@ -1,4 +1,4 @@
-package be.multimedi.mealplanning.mealPlan;
+package be.multimedi.mealplanning.mealSearch;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
