--- src/test/java/be/multimedi/mealplanning/user/RegistrationDatabaseServiceTest.java
@@ -1,16 +1,0 @@
-package be.multimedi.mealplanning.user;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class RegistrationDatabaseServiceTest {
-
-    @Test
-    void registerPotentialUser() {
-    }
-
-    @Test
-    void activateUserAccount() {
-    }
-}
